Update README.md
Modified to reflect a new desire for separation of concerns such that issues here from now on should only be about this reference implementation and issues with the P0267 API or other aspects of P0267 should be filed as issues against that repo. Updated to link to the most recent version of P0267 (currently R4), using wg21.link since that always ensures the most recent version of any WG21 (C++ standardization) paper. Changed Windows requirements to VS 2017 and referenced the need for vcpkg. For those who are curious, this repo currently reflects P0267R2. It will be updated to reflect R4 in the very near future.